refactor: Deck의 컨벤션 수정, 불필요한 메서드 분리 없애기
