+++ src/main/java/blackjack/domain/user/Users.java
@@ -1,8 +1,8 @@
 package blackjack.domain.user;
 
-import java.util.Collections;
-import java.util.LinkedList;
-import java.util.List;
+import blackjack.domain.result.Result;
+
+import java.util.*;
 import java.util.stream.Collectors;
 
 public class Users {
@@ -16,6 +16,12 @@ public class Users {
     public Users(Dealer dealer, List<Player> players) {
         this.dealer = dealer;
         this.players = Collections.unmodifiableList(players);
+    }
+
+    public Map<Player, Result> calculateAllResults() {
+        Map<Player, Result> totalResult = new LinkedHashMap<>();
+        players.forEach(player -> totalResult.put(player, Result.of(dealer, player)));
+        return totalResult;
     }
 
     public List<User> getUsers() {
